Don't bother to download things twice when building from git
When building from git, fetch_tarballs.sh and ooo.lst are present in bootstrap repository, so don't try to download them again

+++ download_external_sources.sh
@@ -3,13 +3,19 @@
 . ./bin/setup
 
 export TARFILE_LOCATION
-GIT_TAG=`echo $OOO_SOURCEDIRNAME | tr "a-z.-" "A-Z__"`
-tmp=`mktemp -q -d`
+export CLONEDIR
 
-cd $tmp
-wget http://cgit.freedesktop.org/ooo-build/bootstrap/plain/fetch_tarballs.sh?id=$GIT_TAG -O fetch_tarballs.sh && chmod 755 fetch_tarballs.sh
-wget http://cgit.freedesktop.org/ooo-build/bootstrap/plain/ooo.lst?id=$GIT_TAG -O ooo.lst
-./fetch_tarballs.sh ooo.lst &&  rm -f fetch_tarballs.sh ooo.lst
-cd ..
-rmdir $tmp
+if test -e $CLONEDIR/bootstrap/fetch_tarballs.sh -a -e $CLONEDIR/bootstrap/ooo.lst; then
+    $CLONEDIR/bootstrap/fetch_tarballs.sh $CLONEDIR/bootstrap/ooo.lst
+else
+    GIT_TAG=`echo $OOO_SOURCEDIRNAME | tr "a-z.-" "A-Z__"`
+    tmp=`mktemp -q -d`
+
+    cd $tmp
+    wget http://cgit.freedesktop.org/ooo-build/bootstrap/plain/fetch_tarballs.sh?id=$GIT_TAG -O fetch_tarballs.sh && chmod 755 fetch_tarballs.sh
+    wget http://cgit.freedesktop.org/ooo-build/bootstrap/plain/ooo.lst?id=$GIT_TAG -O ooo.lst
+    ./fetch_tarballs.sh ooo.lst &&  rm -f fetch_tarballs.sh ooo.lst
+    cd ..
+    rmdir $tmp
+fi
